+++ bin/build-plugin-zip.sh
@@ -70,7 +70,7 @@ status "Installing dependencies... 📦"
 npm cache verify
 npm ci
 status "Generating build... 👷‍♀️"
-npm run build -- --fast
+npm run build -- --skip-types
 
 # Generate the plugin zip file.
 status "Creating archive... 🎁"