+++ packages/ui/src/stack/stories/index.story.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react-webpack5';
+import type { Meta, StoryObj } from '@storybook/react-vite';
 import { Stack } from '../index';
 import { Box } from '../../box';